chore: moved functions to relevant modules
