+++ core/src/crypto.rs
@@ -1,0 +1,7 @@
+use sha2::{Digest, Sha256};
+
+pub fn hash_bytes(data: &[u8]) -> Vec<u8> {
+    let mut hasher = Sha256::new();
+    hasher.update(data);
+    hasher.finalize().to_vec()
+}
