+++ helpers/src/generator.rs
@@ -5,104 +5,19 @@ use cfdkim::{
 };
 use log::{debug, error, info, warn};
 use mailparse::MailHeaderMap;
-use regex_automata::{dfa::regex::Regex as DFARegex, meta::Regex as MetaRegex};
 use slog::{o, Discard, Logger};
-use std::{fs::File, io::Read, path::PathBuf};
+use std::path::PathBuf;
 use trust_dns_resolver::{
     config::{NameServerConfigGroup, ResolverConfig, ResolverOpts},
     TokioAsyncResolver,
 };
-use zkemail_core::{CompiledRegex, Email, EmailWithRegex, PublicKey, RegexInfo, DFA};
+use zkemail_core::{Email, EmailWithRegex, PublicKey, RegexInfo};
 
-use crate::structs::{RegexConfig, RegexPattern};
-
-pub fn read_email_file(path: &PathBuf) -> Result<Vec<u8>> {
-    use std::io::BufReader;
-    let file = File::open(path).map_err(|e| anyhow!("Failed to open email file: {}", e))?;
-    let mut buf_reader = BufReader::new(file);
-    let mut contents = Vec::new();
-    buf_reader
-        .read_to_end(&mut contents)
-        .map_err(|e| anyhow!("Failed to read email contents: {}", e))?;
-    Ok(contents)
-}
-
-pub fn read_regex_config(path: &PathBuf) -> Result<RegexConfig> {
-    let file = File::open(path).map_err(|e| anyhow!("Failed to open regex config file: {}", e))?;
-    let config: RegexConfig =
-        serde_json::from_reader(file).map_err(|e| anyhow!("Failed to read regex config: {}", e))?;
-    Ok(config)
-}
-
-fn create_dfa(re: &DFARegex) -> DFA {
-    let (fwd, fwd_pad) = re.forward().to_bytes_little_endian();
-    let (bwd, bwd_pad) = re.reverse().to_bytes_little_endian();
-    DFA {
-        fwd: fwd[fwd_pad..].to_vec(),
-        bwd: bwd[bwd_pad..].to_vec(),
-    }
-}
-
-fn compile_regex_parts(parts: &[RegexPattern], input: &[u8]) -> Result<Vec<CompiledRegex>> {
-    parts
-        .iter()
-        .map(|part| match part {
-            RegexPattern::Match { pattern } => {
-                let verify_dfa_re = DFARegex::new(pattern)?;
-                if verify_dfa_re.find_iter(input).count() != 1 {
-                    return Err(anyhow!("Input doesn't match regex pattern: {:?}", part));
-                }
-
-                Ok(CompiledRegex {
-                    verify_re: create_dfa(&verify_dfa_re),
-                    capture_str: None,
-                })
-            }
-            RegexPattern::Capture {
-                prefix,
-                capture,
-                suffix,
-            } => {
-                let pattern_dfa = format!("{}{}{}", prefix, capture, suffix);
-                let pattern_meta = format!("({})({})({})", prefix, capture, suffix);
-
-                let verify_dfa_re = DFARegex::new(&pattern_dfa)?;
-                if verify_dfa_re.find_iter(input).count() != 1 {
-                    return Err(anyhow!("Input doesn't match regex pattern: {:?}", part));
-                }
-
-                let verify_meta_re = MetaRegex::new(&pattern_meta)?;
-                let mut caps = verify_meta_re.create_captures();
-                verify_meta_re.captures(input, &mut caps);
-
-                let capture_str = caps
-                    .get_group(2)
-                    .and_then(|capture| String::from_utf8(input[capture.range()].to_vec()).ok())
-                    .ok_or_else(|| anyhow!("Capture contains invalid UTF-8 data"))?;
-
-                Ok(CompiledRegex {
-                    verify_re: create_dfa(&verify_dfa_re),
-                    capture_str: Some(capture_str),
-                })
-            }
-        })
-        .collect()
-}
-
-fn extract_email_body(email: &mailparse::ParsedMail) -> Result<Vec<u8>> {
-    if email.subparts.is_empty() {
-        return email.get_body_raw().map_err(Into::into);
-    }
-
-    email
-        .subparts
-        .iter()
-        .find(|part| part.ctype.mimetype == "text/html")
-        .or_else(|| email.subparts.first())
-        .ok_or_else(|| anyhow!("No valid email body found"))?
-        .get_body_raw()
-        .map_err(Into::into)
-}
+use crate::{
+    email::extract_email_body,
+    file::{read_email_file, read_regex_config},
+    regex::compile_regex_parts,
+};
 
 pub async fn generate_email_inputs(from_domain: &str, email_path: &PathBuf) -> Result<Email> {
     let logger = Logger::root(Discard, o!());
