+++ helpers/src/lib.rs
@@ -1,5 +1,8 @@
-mod input_generators;
+mod email;
+mod file;
+mod generator;
+mod regex;
 mod structs;
 
-pub use input_generators::*;
+pub use generator::{generate_email_inputs, generate_email_with_regex_inputs};
 pub use structs::*;
